+++ src/Azure.Functions.Cli/Kubernetes/KEDA/V1/KedaV1ResourceFactory.cs
@@ -1,32 +1,34 @@
 using System;
 using System.Collections.Generic;
 using System.Linq;
-using Azure.Functions.Cli.Kubernetes.KEDA.V2.Models;
+using Azure.Functions.Cli.Kubernetes.KEDA.V1.Models;
 using Azure.Functions.Cli.Kubernetes.Models;
 using Azure.Functions.Cli.Kubernetes.Models.Kubernetes;
 using Newtonsoft.Json.Linq;
 
-namespace Azure.Functions.Cli.Kubernetes.KEDA.V2
+namespace Azure.Functions.Cli.Kubernetes.KEDA.V1
 {
-    public static class KedaV2Helper
+    public class KedaV1ResourceFactory
     {
         internal static ScaledObjectV1Alpha1 GetScaledObject(string name, string @namespace, TriggersPayload triggers,
-               DeploymentV1Apps deployment, int? pollingInterval, int? cooldownPeriod, int? minReplicas, int? maxReplicas)
+            DeploymentV1Apps deployment, int? pollingInterval, int? cooldownPeriod, int? minReplicas, int? maxReplicas)
         {
-            // TODO: Migrate to new version
             return new ScaledObjectV1Alpha1
             {
                 Metadata = new ObjectMetadataV1
                 {
                     Name = name,
                     Namespace = @namespace,
-                    Labels = new Dictionary<string, string>()
+                    Labels = new Dictionary<string, string>
+                    {
+                        {"deploymentName", deployment.Metadata.Name}
+                    }
                 },
                 Spec = new ScaledObjectSpecV1Alpha1
                 {
                     ScaleTargetRef = new ScaledObjectScaleTargetRefV1Alpha1
                     {
-                        Name = deployment.Metadata.Name
+                        DeploymentName = deployment.Metadata.Name
                     },
                     PollingInterval = pollingInterval,
                     CooldownPeriod = cooldownPeriod,
@@ -56,7 +58,6 @@ namespace Azure.Functions.Cli.Kubernetes.KEDA.V2
                 .Where(i => i.Value.Type == JTokenType.String)
                 .ToDictionary(k => k.Key, v => v.Value.ToString());
 
-            // TODO: Check for breaking changes
             if (t["type"].ToString().Equals("rabbitMQTrigger", StringComparison.InvariantCultureIgnoreCase))
             {
                 metadata["host"] = metadata["connectionStringSetting"];
@@ -66,7 +67,7 @@ namespace Azure.Functions.Cli.Kubernetes.KEDA.V2
             return metadata;
         }
 
-        internal static string GetKedaTrigger(string triggerType)
+        private static string GetKedaTrigger(string triggerType)
         {
             if (string.IsNullOrEmpty(triggerType))
             {
@@ -75,7 +76,6 @@ namespace Azure.Functions.Cli.Kubernetes.KEDA.V2
 
             triggerType = triggerType.ToLower();
 
-            // TODO: Check if we need to add new ones
             switch (triggerType)
             {
                 case "queuetrigger":
